# [REVIEW] Optimize has_duplicate_edges 