--- a/python/cugraph/cugraph/dask/structure/mg_property_graph.py
+++ b/python/cugraph/cugraph/dask/structure/mg_property_graph.py
@@ -654,10 +654,10 @@ class EXPERIMENTAL__MGPropertyGraph:
         if len(df.columns) == 0:
             return False
 
-        unique_pair_len = df.drop_duplicates(split_out=df.npartitions).size
+        unique_pair_len = df.drop_duplicates(split_out=df.npartitions).shape[0]
         # if unique_pairs == len(df)
         # then no duplicate edges
-        return unique_pair_len != df.size
+        return unique_pair_len != df.shape[0]
 
     def __create_property_lookup_table(self, edge_prop_df):
         """