@VibhuJawa 2022-07-15
> Just FYI: I am not sure this is relevant here, but cuGraph C++ graph_view_t has count_multi_edges() (https://github.com/rapidsai/cugraph/blob/branch-22.08/cpp/include/cugraph/graph_view.hpp#L680).

@seunghwak , Thanks for the input , raised an issue here to track it. https://github.com/rapidsai/cugraph/issues/2417